Update CarInsuranceQuoteDrill - a quote generator class was added
Here I added a QuoteGenereator class which has a method that generates a quote based on specific conditions. I also added a way for that quote to get calculated and added to a property in that current object before a record (created from that object) gets stored on the database.